Refacto OrderAmountUpdater to handle the cart rules update
